+++ relay-server/src/processing/utils/mod.rs
@@ -4,4 +4,3 @@ pub mod dynamic_sampling;
 pub mod event;
 #[cfg(feature = "processing")]
 pub mod store;
-pub mod transaction;
